emacs: changes to font customization
Don't bother trying to change the font when running without a GUI.

Try fonts from most-preferred to least-preferred, stopping once the font
was set successfully.  Previously, fonts were tried from least-preferred
to most-preferred, so that the most-preferred available font "won".  The
old approach meant set-frame-font could be called successfully multiple
times, which was inefficient.

Call set-frame-font with INHIBIT-CUSTOMIZE as t.  The parameters to the
my-set-fonts invocation take care of persisting the font choice, so it
doesn't need to be saved by Customize.  (This change is a no-op in the
context of my init.el, which disables Customize elsewhere; but in
isolation it's an improvement.)

Refactor so that there is a common list of fonts and an OS-specific list
of additional fonts.

Changes to font list:
- Add "Go Mono"
- Add "Cascadia Code" (default Windows Terminal font)
- Reduce "Hack" font size