emacs: sidebar tweaks
Don't move the cursor to the sidebar when it's opened.

Don't delete the dired-sidebar when delete-other-windows is run.  Note
that ibuffer-sidebar already behaves this way.

Enhance (?) my-sidebar-toggle so that it doesn't enable one sidebar
while disabling the other.  Instead, if one sidebar is already disabled,
the enabled sidebar is also disabled.