cwm: move bemenu to the bottom
bemenu is appearing above the top of the screen when there are multiple
monitors with differing display resolutions and it is invoked from the
monitor with the shorter resolution.

When moved to the "bottom", bemenu is always visible, although on the
monitor with the short resolution, it's appearing in the middle of the
screen rather than at the bottom of it.